Move FunctionCreationStatus out of _output module
